fix other errors

+++ budget-optimizer-backend/src/main/java/com/budgetoptimizer/budget_optimizer_backend/service/UsuarioService.java
@@ -390,7 +390,7 @@ public class UsuarioService {
 
                 // Beneficios
                 .descuentoPorcentaje(
-                        BigDecimal.valueOf(accountType.getDescuentoPorcentaje())
+                        accountType.getDescuentoPorcentaje()
                 )
 
                 .tieneBeneficios(
